Changing package versions for net8.0
The `Gridify.EntityFramework.csproj` file has changed the version of the `Microsoft.EntityFrameworkCore` package for the `net8.0` target platform. Previously the version was specified as `[8.0.0,9.0.0)`, now it has been changed to `[8.0.0,10.0.0)`.

In the `Gridify.csproj` file, the version of the `Microsoft.Extensions.DependencyInjection.Abstractions` package for the `net8.0` target platform has been changed. Previously the version was specified as `[8.0.0,9.0.0)`, now it has been changed to `[8.0.0,10.0.0)`.

#240